# doc/api.yml
openapi: '3.0.2'
info:
  title: Clubs HEI
  version: '1.0'
servers:
  - url: https://api.server.test/v1
tags:
  - name: Students
  - name: Groups
  - name: Clubs
  - name: Events
paths:
  /students:
    get:
      tags:
        - Students
      operationId: getStudents
      summary: Get all students
      description: Get all students with details
      parameters:
        - name: first_name
          in: query
          schema:
            type: string
          required: false
        - name: groups_name
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '400':
          description: Bad request
    put:
      tags:
        - Students
      operationId: Change informations
      summary: Change student's informations
      description: Change student's informations
      parameters:
        - name: groups_id
          in: query
          schema:
            type: integer
          required: true
        - name: club_id
          in: query
          schema:
            type: integer
          required: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudent'
        required: true
      responses:
        '200':
          description: Update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Bad request

  /students/{id}:
    get:
      tags:
        - Students
      operationId: getStudentById
      summary: Get student by id
      description: Get student by id with details
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Details of one student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Bad request
    patch:
      tags:
        - Students
      operationId: Change one or more information
      summary: Change one or more student's information
      description: |
        Changes options : 
        * groups_id: change
        * club_id: change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: groups_id
          in: query
          required: false
          schema:
            type: integer
        - name: club_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        204:
          description: update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        '400':
          description: Bad request  

  /members:
    get:
      tags:
        - Students
      operationId: getStudentsClubNotNull
      summary: Get all students with club not null
      description: Get all students with club not null
      parameters:
        - name: club_name
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: List of students with club not null
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '400':
          description: Bad request  

  /groups:
    get:
      tags:
        - Groups
      operationId: getGroups
      summary: Get all groups
      description: Get all groups with details
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
    put:
      tags:
        - Groups
      operationId: Change informations
      summary: Change group's informations
      description: Change group's informations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
        required: true
      responses:
        '200':
          description: Update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad request

  /groups/{id}:
    get:
      tags:
        - Groups
      operationId: getGroupById
      summary: Get group by id
      description: Get group by id with details
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Details of one group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad request   
    patch:
      tags:
        - Groups
      operationId: Change one information
      summary: Change one or more group's information
      description: |
        Changes options : 
        * name: change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        204:
          description: update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        '400':
          description: Bad request               

  /clubs:
    get:
      tags:
        - Clubs
      operationId: getClubs
      summary: Get all clubs
      description: Get all clubs with details
      parameters:
        - name: name
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: List of clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Club'
        '400':
          description: Bad request
    put:
      tags:
        - Clubs
      operationId: Change information
      summary: Change club's information
      description: Change club's information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClub'
        required: true
      responses:
        '200':
          description: Update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          description: Bad request

  /clubs/{id}:
    get:
      tags:
        - Clubs
      operationId: getClubById
      summary: Get club by id
      description: Get club by id with details
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Details of one club
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Club'
        '400':
          description: Bad request    
    patch:
      tags:
        - Clubs
      operationId: Change one or more information
      summary: Change one or more club's information
      description: |
        Changes options : 
        * name: change
        * description: change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: description
          in: query
          required: false
          schema:
            type: string
      responses:
        204:
          description: update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        '400':
          description: Bad request             

  /events:
    get:
      tags:
        - Events
      operationId: getEvents
      summary: Get all events
      description: Get all events with details
      parameters:
        - name: event_name
          in: query
          schema:
            type: string
          required: false
        - name: club_name
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: List of Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
    put:
      tags:
        - Events
      operationId: Change information
      summary: Change event's information
      description: Change event's information
      parameters:
        - name: club_id
          in: query
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvent'
        required: true
      responses:
        '200':
          description: Update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad request

  /events/{id}:
    get:
      tags:
        - Events
      operationId: getEventById
      summary: Get event by id
      description: Get event by id with details
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Details of one event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
    patch:
      tags:
        - Events
      operationId: Change one or more information
      summary: Change one or more event's information
      description: |
        Changes options : 
        * start_event: change
        * end_event: change
        * club_id: change
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: start_event
          in: query
          required: false
          schema:
            type: string
            format: LocalDateTime
        - name: end_event
          in: query
          required: false
          schema:
            type: string
            format: LocalDateTime
        - name: club_id
          in: query
          required: false
          schema:
            type: integer
      responses:
        204:
          description: update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        '400':
          description: Bad request         


components:
  schemas:
    CreateStudent:
      type: object
      properties:
        ref:
          type: string
          example: STD21001
        firstName:
          type: string
          example: Rabe
        lastName:
          type: string
          example: Rakoto
        email:  
          type: string
          example: hei.rakoto@gmail.com
        gender:
          type: string
          enum:
            - F
            - M
    Student:
      allOf:
        - $ref: "#/components/schemas/CreateStudent"
      properties:
        group:
         $ref: '#/components/schemas/Group'
        clubs:
          type: array
          items:
            $ref: '#/components/schemas/Club'
        id:
          type: integer
          example: 10     
      
    CreateGroup:
      type: object
      properties:
        name:
          type: string
          example: G1
    Group:
      allOf:
        - $ref: '#/components/schemas/CreateGroup'
      properties:
        id:
          type: integer
          example: 10

    CreateClub:
      type: object
      properties:
        name:
          type: string
          example: basket
        description:
          type: string
          example: string
    Club:
      allOf:
        - $ref: '#/components/schemas/CreateClub'
      properties:
        creation_date:
          type: string
          format: date
          example: '26-08-2022'
        id:
          type: integer
          example: 10  

    CreateEvent:
      type: object
      properties:
        name:
          type: string
          example: assistance
        description:
          type: string
          example: string
        startEvent:
          type: string
          format: LocalDateTime
          example: '2022-01-02T08:00:00'
        endEvent:
          type: string
          format: LocalDateTime
          example: '2022-01-02T10:00:00'
    Event:
      allOf:
        - $ref: '#/components/schemas/CreateEvent'
      properties:
        club:
          $ref: '#/components/schemas/Club'
        id:
          type: integer
          example: 10